Split tests into separate files
Especially to clean up snapshot files
