+++ tests/unit-tests/ambiguous/semi-false.test.ts
@@ -1,0 +1,18 @@
+import { describe } from 'vitest';
+
+import {
+  ambiguousCases,
+  ambiguousExpressionTest,
+} from '../../helpers/ambiguous';
+import { describeConfig } from '../../helpers/config';
+
+describe('ambiguous', () => {
+  describeConfig(
+    {
+      name: 'semi: false',
+      options: { semi: false },
+    },
+    ambiguousExpressionTest,
+    ambiguousCases
+  );
+});
