+++ tests/unit-tests/config/semi-false.test.ts
@@ -1,0 +1,3 @@
+import { describeConfig } from '../../helpers/config';
+
+describeConfig({ name: 'semi: false', options: { semi: false } });
