+++ tests/unit-tests/format.test.ts
@@ -1,0 +1,6 @@
+import { describe } from 'vitest';
+import { describeConfig } from '../helpers/config';
+
+describe('format', () => {
+  describeConfig({ name: 'default' });
+});
